Julius_AI/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
google-generative-aipython-dotenv